fix: Remove temporary cleanup tools and fix Edge Runtime warnings
- Remove "Fix Products" and "Delete All Products" buttons from seller dashboard
- Delete cleanup API endpoint that's no longer needed
- Fix Edge Runtime warnings by splitting auth configuration
- Create edge-compatible auth for middleware without bcrypt
- Maintain all security checks while improving build performance
