+++ lib/auth.ts
@@ -8,8 +8,10 @@ import {
   getClientIP,
   validateEmail,
 } from '@/lib/auth-security';
+import { authConfig } from '@/lib/auth-config';
 
 export const { handlers, auth, signIn, signOut } = NextAuth({
+  ...authConfig,
   providers: [
     CredentialsProvider({
       name: 'credentials',
@@ -91,36 +93,4 @@ export const { handlers, auth, signIn, signOut } = NextAuth({
       },
     }),
   ],
-  session: {
-    strategy: 'jwt',
-    maxAge: 24 * 60 * 60, // 24 hours session timeout
-    updateAge: 60 * 60, // Update session every 1 hour if user is active
-  },
-  callbacks: {
-    async jwt({ token, user }) {
-      if (user) {
-        token.role = user.role;
-        token.sellerProfile = user.sellerProfile;
-      }
-      return token;
-    },
-    async session({ session, token }) {
-      if (session.user) {
-        session.user.id = token.sub!;
-        session.user.role = token.role as string;
-        session.user.sellerProfile = token.sellerProfile as {
-          id: string;
-          shopName: string;
-          displayName: string;
-          bio?: string | null;
-          region?: string | null;
-          avatarUrl?: string | null;
-        } | null;
-      }
-      return session;
-    },
-  },
-  pages: {
-    signIn: '/auth/login',
-  },
 });
