+++ middleware.ts
@@ -1,6 +1,6 @@
 import createMiddleware from 'next-intl/middleware';
 import { NextRequest, NextResponse } from 'next/server';
-import { auth } from '@/lib/auth';
+import { auth } from '@/lib/auth-edge';
 
 const intlMiddleware = createMiddleware({
   locales: ['fa', 'en'],
